refactor(persistence): split database.rs into focused submodules (TD-024)
Behavior-preserving split of parish-persistence database.rs (~1430 LOC)
into four focused submodules under database/:
- schema.rs: DDL, WAL setup, migrate, lock_recovered
- branches.rs: BranchInfo, branch CRUD, row mapping
- journal.rs: SnapshotInfo, snapshot + journal ops, compaction
- async_adapter.rs: AsyncDatabase Tokio spawn_blocking wrapper

database/mod.rs acts as the Rust-2018 hub, re-exporting all public
types so external use-paths are unchanged. All tests preserved in
mod.rs. Part of #1200 (TD-024).
